requirements.txt
python-pptx==0.6.21
anthropic==0.3.11
requests==2.31.0
tenacity==8.2.3
python-dotenv==1.0.0
pdf2image==1.16.3
lxml==4.9.3
Pillow==10.0.0